Switch val1 & val2 to be integers.
These two values were originally long values, due to the use of strtoul, but even that wasn't correct, as they should have been 'unsigned long' but that would have caused problems.  Now things are back to the right sizes and aren't causing possible silly bugs due to pointers to their address and wrong int/long values.

+++ src/commands.c
@@ -124,8 +124,8 @@ int toUint(const char *value, int base, int maxValue) {
  * Docker ADD and COPY instructions.
  */
 int runCommands() {
-    long val1 = 0;
-    long val2 = 0;
+    int val1 = 0;
+    int val2 = 0;
     int err = 0;
     int errCount = 0;
     int cmd = CMD_ERR;
